Use uint to match types of surround code instead of stray size_t
GIT_SILENT

+++ kdevplatform/serialization/tests/test_itemrepository.cpp
@@ -61,7 +61,7 @@ struct TestItemRequest
     }
 
     //Should return the size of an item created with createItem
-    size_t itemSize() const
+    uint itemSize() const
     {
         return m_item.itemSize();
     }
